Project 6th Sem/backend/requirements.txt
# Core
Flask==3.0.2
python-gvm==24.6.0
python-nmap==0.7.1
pandas==2.2.2
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.3.2
gunicorn==21.2.0
waitress==2.1.2  # For Windows compatibility